renaming log to logs due to window build issues
